// src/types.ts
export interface ITransaction {
  FirstName: string;
  LastName: string;
  Gender: string;
  Latitude: number;
  Longitude: number;
  CreditCardNumber: string;
  CreditCardType: string;
  Email: string;
  DomainName: string;
  PhoneNumber: string;
  MacAddress: string;
  URL: string;
  UserName: string;
  LastLogin: string;
  PaymentMethod: string;
}